# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "Node.js API model",
    "description": "Node.js Express API Microservice with Mongo",
    "contact": {
      "email": "david.a.m@live.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8080/api"
    },
    {
      "url": "https://localhost:8080/api"
    }
  ],
...
  "paths": {
...
    "/auth/google": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login with Google account",
        "description": "Login with a Google account",
        "operationId": "googleSignIn",
        "requestBody": {
          "description": "Log in using Google Auth",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginAuthorized"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get all categories",
        "description": "Get all categories",
        "operationId": "getCategories",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No categories found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Add a new category to the store",
        "description": "Add a new category to the store",
        "operationId": "postCategories",
        "requestBody": {
          "description": "Create a new category in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCategory"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Category created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get category by ID",
        "description": "Find a category searching by ID",
        "operationId": "getCategoryById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The MongoID of the category to be fetched",
            "required": true,
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Delete an existing category",
        "description": "Delete an existing category by Id",
        "operationId": "deleteCategories",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The MongoID of the category to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid category ID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Category ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      },
      "put": {
        "tags": ["Categories"],
        "summary": "Update an existing category",
        "description": "Update an existing category by Id",
        "operationId": "updateCategories",
        "requestBody": {
          "description": "Update an existent category in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCategory"
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The MongoID of the category to be fetched",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid MongoID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token, user unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "There is no category with that MongoID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "token_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Category": {
        "required": ["name", "category"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24,
            "example": "6208136134ffbeeabc4b1ead"
          },
          "name": {
            "type": "string",
            "example": "PASTRIES"
          },
          "status": {
            "type": "boolean",
            "description": "current status in the store"
          }
        }
      },
      "GoogleLogin": {
        "required": ["id_token"],
        "type": "object",
        "properties": {
          "id_token": {
            "type": "string",
            "example": "zndksdfjvrEEASJFu38738dhfnFDS"
          }
        }
      },
...
      "LoginAuthorized": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiI2MjAzMjI1MGZkY2E1Y2ExYTE3MGQ5NzciLCJpYXQiOjE2NjMzNDI1OTMsImV4cCI6MTY2MzM1Njk5M30.xFPKG8n_TIfPGRI6rlpSE3KSNyOC7BTOnYOqNnJtzcw"
          }
        }
      },
      "NewCategory": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "PASTRY ROLL"
          }
        }
      },
...
      "User": {
        "required": ["name", "category"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 24,
            "maxLength": 24,
            "example": "68373827924ffbeeabc4b1ed"
          },
          "name": {
            "type": "string",
            "example": "John Smith"
          },
          "email": {
            "type": "string",
            "example": "test3@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "12345678"
          },
          "img": {
            "type": "string",
            "example": "test.jpg"
          },
          "role": {
            "type": "string",
            "example": "ADMIN_ROLE",
            "enum": ["ADMIN_ROLE", "USER_ROLE"]
          },
          "google": {
            "type": "boolean",
            "description": "has Google authentication"
          },
          "status": {
            "type": "boolean",
            "description": "current status in the store"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "BadRequest"
          },
          "title": {
            "type": "string",
            "example": "Invalid ID"
          },
          "detail": {
            "type": "string",
            "example": "The ID supplied is not a valid MongoID value"
          }
        }
      }
    },
    "securitySchemes": {
      "token_auth": {
        "type": "apiKey",
        "name": "x-token",
        "in": "header"
      }
    }
  }
}